Refactor all separate todo components/services
Moved all of the separate todo components and services
into one folder called `todos`.
