+++ client/src/app/app-routing.module.ts
@@ -3,14 +3,16 @@ import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { UserListComponent } from './users/user-list.component';
 import { UserProfileComponent } from './users/user-profile.component';
-import { TodosComponent } from './todos/todos.component';
+import { TodoProfileComponent } from './todos/todo-profile.component';
+import { TodoListComponent } from './todos/todo-list.component';
 
 
 const routes: Routes = [
   {path: '', component: HomeComponent},
   {path: 'users', component: UserListComponent},
   {path: 'users/:id', component: UserProfileComponent},
-  {path: 'todos', component: TodosComponent}
+  {path: 'todos', component: TodoListComponent},
+  {path: 'todos/:id', component: TodoProfileComponent}
 ];
 
 @NgModule({
