+++ client/src/app/app.module.ts
@@ -31,9 +31,10 @@ import { HttpClientModule } from '@angular/common/http';
 import { LayoutModule } from '@angular/cdk/layout';
 import { UserCardComponent } from './users/user-card.component';
 import { UserProfileComponent } from './users/user-profile.component';
-import { TodoCardComponent } from './todo-card/todo-card.component';
-import { TodoListComponent } from './todo-list/todo-list.component';
-import { TodoProfileComponent } from './todo-profile/todo-profile.component';
+import { TodoCardComponent } from './todos/todo-card.component';
+import { TodoListComponent } from './todos/todo-list.component';
+import { TodoProfileComponent } from './todos/todo-profile.component';
+import { TodoService } from './todos/todo.service';
 
 const MATERIAL_MODULES: any[] = [
   MatListModule,
@@ -76,7 +77,8 @@ const MATERIAL_MODULES: any[] = [
     LayoutModule,
   ],
   providers: [
-    UserService
+    UserService,
+    TodoService
   ],
   bootstrap: [AppComponent]
 })
