+++ client/src/app/todos/todo-list.component.html
@@ -2,7 +2,7 @@
   <div fxFlex fxFlex.gt-sm="80" fxFlexOffset.gt-sm="10">
 
     <mat-card class="search-card">
-      <mat-card-title class="todo-list-title">Users</mat-card-title>
+      <mat-card-title class="todo-list-title">Todos</mat-card-title>
       <mat-card-content fxLayout="column" >
 
         <div fxLayout="row wrap" fxLayoutGap="10px">
