+++ client/src/app/todos/todo-profile.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute  } from '@angular/router';
 import { Todo } from './todo';
-import { TodoService } from '../todo.service';
+import { TodoService } from './todo.service';
 
 @Component({
   selector: 'app-todo-profile',
