+++ client/src/app/todos/todo.service.spec.ts
@@ -1,5 +1,7 @@
+import { HttpClient } from '@angular/common/http';
+import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
-
+import { Todo } from './todo';
 import { TodoService } from './todo.service';
 
 describe('TodoService', () => {
