+++ client/src/app/todos/todo.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { environment } from '../environments/environment';
+import { environment } from '../../environments/environment';
 import { Observable } from 'rxjs';
-import { Todo } from './todo-profile/todo';
+import { Todo } from './todo';
 
 
 @Injectable()
